Add Spring Boot compatability check workflow
Adds a GHA workflow to check Spring Boot compatability of the framework.

The workflow exercises the samples and dynamically sets the version of
each samples Spring Boot plugin.

We do the dynamic version setting using the power of Gradle init scripts
and the fact that each of the samples is exposed as a Gradle module.
Doing this has a positive side effect that our samples are kept up to date
for both Maven and Gradle.

Reworks the Gradle aspect of the samples as follows:

- removes settings.xml in each sample in favor of root samples/setting.xml
- removes Gradle wrapper from each sample in favor of root samples/gradlew
- cleans up build.gradle (they were missing dependencies to work)
- adds README-GRADLE.md at samples/

Signed-off-by: Chris Bono <chris.bono@gmail.com>