Fix docs.yml workflow (add 'package' command)
The recent move to do config props generation only at docs
generation time left the docs.yml broken as it requires the
config props generator class to be compiled prior to running.

This updates the './build-docs.sh' to include the package command
as well as update the docs.yml to use the same script.

Signed-off-by: Chris Bono <chris.bono@gmail.com>